+++ Forge/src/main/java/net/merchantpug/apugli/ApugliForgeEventHandler.java
@@ -314,7 +314,6 @@ public class ApugliForgeEventHandler {
 
     @SubscribeEvent
     public static void prePowerLoad(AddReloadListenerEvent event) {
-        Apugli.clearErrorLoggedDataSet();
         TextureUtil.getCache().clear();
     }